Working game

+++ cannon/src/main/java/cannongame/DegreeConversion.java
@@ -2,7 +2,7 @@ package cannongame;
 
 public class DegreeConversion implements IDegreeConversion {
 
-    public double ConvertToDegrees(int angle) {
+    public double ConvertToDegrees(double angle) {
         double result = (angle * (Math.PI / 180));
         return result;
     }
