+++ cannon/src/main/java/cannongame/IAngleFlow.java
@@ -1,5 +1,5 @@
 package cannongame;
 
 public interface IAngleFlow {
-    int getAngle();
+    double getAngle();
 }
