+++ cannon/src/main/java/cannongame/IDegreeConversion.java
@@ -2,6 +2,6 @@ package cannongame;
 
 public interface IDegreeConversion {
 
-    double ConvertToDegrees(int angle);
+    double ConvertToDegrees(double angle);
 
 }
